PythonProgram-oneToFifty
NEW Data Analytics Bootcamp Python Project

One of the languages that we pay special attention to in our apprenticeship programs is Python. Python can be used to automate DevOps and routine day to day activities in the AWS cloud environment.

Therefore, prior to your interview, we'd like you to write a simple Python program to demonstrate your technical and problem-solving capabilities. For those with Python experience, we anticipate that this should be fairly straightforward. For those without prior experience, reference our resources section below to help you complete this project.

The Project

Write a python program that iterates the integers from 1 to 50.
For multiples of three print "Cloud" instead of the number
For multiples of seven print "Computing"
For numbers which are multiples of both three and seven print "CloudComputing"